추가 분할 컴파일
